extract menu
